requirements.txt
fastapi==0.77.1
uvicorn==0.17.6
gunicorn==20.1.0
pydantic==1.9.1
psycopg2-binary~=2.9.3
aioredis==2.0.1
elasticsearch[async]==7.9.1
orjson==3.6.8
python-dotenv==0.20.0

pytest~=7.1.1
pytest-asyncio
dataclasses~=0.6
multidict~=6.0.2
aiohttp~=3.8.1